pyproject.toml
[tool.poetry]
name = "obar"
version = "0.1.0"
description = ""
authors = ["ali <alysfryly075@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
pytest-django = "^4.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
